Fix incorrect casting used in the xcb platform
They cast ints to void *, so instead make those go via intptr_t.

(cherry picked from commit 149682e8ceaf6048be6a82bda8397a3c87a45da2)

+++ src/plugins/platforms/xcb/qxcbnativeinterface.cpp
@@ -170,13 +170,15 @@ void *QXcbNativeInterface::nativeResourceForScreen(const QByteArray &resourceStr
         result = appUserTime(xcbScreen);
         break;
     case ScreenHintStyle:
-        result = reinterpret_cast<void *>(xcbScreen->hintStyle() + 1);
+        result = reinterpret_cast<void *>(static_cast<std::intptr_t>(xcbScreen->hintStyle() + 1));
         break;
     case ScreenSubpixelType:
-        result = reinterpret_cast<void *>(xcbScreen->subpixelType() + 1);
+        result =
+                reinterpret_cast<void *>(static_cast<std::intptr_t>(xcbScreen->subpixelType() + 1));
         break;
     case ScreenAntialiasingEnabled:
-        result = reinterpret_cast<void *>(xcbScreen->antialiasingEnabled() + 1);
+        result = reinterpret_cast<void *>(
+                static_cast<std::intptr_t>(xcbScreen->antialiasingEnabled() + 1));
         break;
     case TrayWindow:
         if (QXcbSystemTrayTracker *s = systemTrayTracker(screen))
@@ -186,7 +188,7 @@ void *QXcbNativeInterface::nativeResourceForScreen(const QByteArray &resourceStr
         result = getTimestamp(xcbScreen);
         break;
     case RootWindow:
-        result = reinterpret_cast<void *>(xcbScreen->root());
+        result = reinterpret_cast<void *>(static_cast<std::intptr_t>(xcbScreen->root()));
         break;
     case CompositingEnabled:
         if (QXcbVirtualDesktop *vd = xcbScreen->virtualDesktop())
@@ -373,7 +375,8 @@ void *QXcbNativeInterface::x11Screen()
     QXcbIntegration *integration = QXcbIntegration::instance();
     QXcbConnection *defaultConnection = integration->defaultConnection();
     if (defaultConnection)
-        return reinterpret_cast<void *>(defaultConnection->primaryScreenNumber());
+        return reinterpret_cast<void *>(
+                static_cast<std::intptr_t>(defaultConnection->primaryScreenNumber()));
     return nullptr;
 }
 
@@ -382,7 +385,8 @@ void *QXcbNativeInterface::rootWindow()
     QXcbIntegration *integration = QXcbIntegration::instance();
     QXcbConnection *defaultConnection = integration->defaultConnection();
     if (defaultConnection)
-        return reinterpret_cast<void *>(defaultConnection->rootWindow());
+        return reinterpret_cast<void *>(
+                static_cast<std::intptr_t>(defaultConnection->rootWindow()));
     return nullptr;
 }
